renamed StructureModifier #547
