+++ analysis/filter/StructureModifier/src/main/kotlin/de/maibornwolff/codecharta/filter/structuremodifier/FolderMover.kt
@@ -1,4 +1,4 @@
-package de.maibornwolff.codecharta.filter.structurechanger
+package de.maibornwolff.codecharta.filter.structuremodifier
 
 import de.maibornwolff.codecharta.model.*
 import mu.KotlinLogging
