+++ analysis/filter/StructureModifier/src/test/kotlin/de/maibornwolff/codecharta/filter/structuremodifier/ProjectStructurePrinterTest.kt
@@ -1,4 +1,4 @@
-package de.maibornwolff.codecharta.filter.structurechanger
+package de.maibornwolff.codecharta.filter.structuremodifier
 
 import de.maibornwolff.codecharta.model.Project
 import de.maibornwolff.codecharta.serialization.ProjectDeserializer
