+++ analysis/filter/StructureModifier/src/test/resources/sample_project.cc.json
